requirements.txt
browser-use
langchain-google-genai
python-dotenv
pydantic
PyPDF2 